+++ src/filterPOAndWriteTranslateSync.js
@@ -10,7 +10,9 @@ import toObjectBy from 'to-object-by';
 import readAllMessageAsObjectSync from './readAllMessageAsObjectSync';
 import readAllPOAsObjectSync from './readAllPOAsObjectSync';
 
-function filterPOAndWriteTranslateSync(srcPatterns, { messagesPattern, output, multi = false }) {
+const isAJSONFile = string => /.json/.test(string);
+
+function filterPOAndWriteTranslateSync(srcPatterns, { messagesPattern, output }) {
   const translationTable = readAllPOAsObjectSync(srcPatterns);
   const messageList = flowRight(
     flatten,                    // 3. return flatten object values
@@ -25,8 +27,12 @@ function filterPOAndWriteTranslateSync(srcPatterns, { messagesPattern, output, m
     })),
   }));
 
-  if (multi) {
-    mkdirp.sync(output);
+  if (isAJSONFile(output)) {
+    mkdirp.sync(path.dirname(output)); // ensure the output folder exists
+    fs.writeFileSync(output, JSON.stringify(result, null, 0));
+    console.log(chalk.green(`> [react-intl-po] write file -> ${output} ✔️\n`));
+  } else {
+    mkdirp.sync(output); // ensure the output folder exists
 
     Object.keys(result).map(lang => {
       fs.writeFileSync(path.join(output, `${lang}.json`), JSON.stringify(result[lang], null, 0));
@@ -34,10 +40,6 @@ function filterPOAndWriteTranslateSync(srcPatterns, { messagesPattern, output, m
         `> [react-intl-po] write file -> ${path.join(output, `${lang}.json`)} ✔️`));
       return null;
     });
-  } else {
-    mkdirp.sync(path.dirname(output)); // ensure the output folder exists
-    fs.writeFileSync(output, JSON.stringify(result, null, 0));
-    console.log(chalk.green(`> [react-intl-po] write file -> ${output} ✔️\n`));
   }
 }